+++ lib/s_connectivity/s_connectivity.dart
@@ -3,4 +3,4 @@
 /// This library exports all public APIs for the s_connectivity package.
 library;
 
-export 'src/s_connection.dart';
+export 'src/s_connectivity.dart';